Only enable IPv6 forwarding if IPv6 is enabled

+++ calicoctl/calico_ctl/node.py
@@ -562,10 +562,11 @@ def _setup_ip_forwarding():
         sys.exit(1)
 
     try:
-        with open('/proc/sys/net/ipv6/conf/all/forwarding', 'w') as f:
-            f.write("1")
+        if os.path.exists('/proc/sys/net/ipv6'):
+            with open('/proc/sys/net/ipv6/conf/all/forwarding', 'w') as f:
+                f.write("1")
     except:
-        print "ERROR: Could not enable ipv4 forwarding."
+        print "ERROR: Could not enable ipv6 forwarding."
         sys.exit(1)
